# back/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API documentation for Easy DEVS.",
		"title": "Easy DEVS API",
		"contact": {},
		"version": "1.0"
	},
	"host": "localhost:3000",
	"basePath": "/",
	"paths": {
...
		"/user": {
			"get": {
				"description": "Retrieve a list of all user",
				"produces": ["application/json"],
				"tags": ["user"],
				"summary": "Get all user",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/model.User"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/user/{id}": {
			"get": {
				"description": "Retrieve a single user by their ID",
				"produces": ["application/json"],
				"tags": ["user"],
				"summary": "Get a user by ID",
				"parameters": [
					{
						"type": "string",
						"description": "User ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/model.User"
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			},
			"delete": {
				"description": "Delete an existing user by their ID",
				"tags": ["user"],
				"summary": "Delete a user by ID",
				"parameters": [
					{
						"type": "string",
						"description": "User ID",
						"name": "id",
						"in": "path",
						"required": true
					},
					{
						"description": "User password confirmation",
						"name": "user",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/request.PasswordRequest"
						}
					}
				],
				"responses": {
					"204": {
						"description": "No Content",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			},
			"patch": {
				"description": "Update an existing user with partial data",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["user"],
				"summary": "Update a user",
				"parameters": [
					{
						"type": "string",
						"description": "User ID",
						"name": "id",
						"in": "path",
						"required": true
					},
					{
						"description": "Partial user update",
						"name": "user",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/request.UpdateUserRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/model.User"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/workspace": {
			"get": {
				"description": "Retrieve a list of all workspace",
				"produces": ["application/json"],
				"tags": ["workspace"],
				"summary": "Get all workspace",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/model.Workspace"
							}
						}
					}
				}
			},
			"post": {
				"description": "Create a new workspace and store it in the database",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["workspace"],
				"summary": "Create a new workspace",
				"parameters": [
					{
						"description": "Workspace object",
						"name": "workspace",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/model.Workspace"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/model.Workspace"
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		},
		"/workspace/{id}": {
			"get": {
				"description": "Retrieve a single workspace by its ID",
				"produces": ["application/json"],
				"tags": ["workspace"],
				"summary": "Get a workspace by ID",
				"parameters": [
					{
						"type": "string",
						"description": "Workspace ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/model.Workspace"
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			},
			"delete": {
				"description": "Delete an existing workspace by its ID",
				"tags": ["workspace"],
				"summary": "Delete a workspace by ID",
				"parameters": [
					{
						"type": "string",
						"description": "Workspace ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"204": {
						"description": "No Content",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			},
			"patch": {
				"description": "Update an existing workspace with partial data",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["workspace"],
				"summary": "Update a workspace",
				"parameters": [
					{
						"type": "string",
						"description": "Workspace ID",
						"name": "id",
						"in": "path",
						"required": true
					},
					{
						"description": "Partial workspace update",
						"name": "workspace",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/model.Workspace"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					}
				}
			}
		}
	},
	"definitions": {
...
		"model.User": {
			"type": "object",
			"required": ["email", "password", "username"],
			"properties": {
				"createdAt": {
					"type": "string"
				},
				"deletedAt": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"fullname": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"password": {
					"type": "string",
					"maxLength": 50,
					"minLength": 6
				},
				"refresh_token": {
					"type": "string"
				},
				"updatedAt": {
					"type": "string"
				},
				"username": {
					"type": "string",
					"maxLength": 50,
					"minLength": 3
				}
			}
		},
		"model.Workspace": {
			"type": "object",
			"properties": {
				"createdAt": {
					"type": "string"
				},
				"deletedAt": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"updatedAt": {
					"type": "string"
				}
			}
		},
...
		"request.PasswordRequest": {
			"type": "object",
			"properties": {
				"password": {
					"type": "string"
				}
			}
		},
...
		"request.UpdateUserRequest": {
			"type": "object",
			"properties": {
				"names": {
					"type": "string"
				}
			}
		},
...
	}
}